requirements.txt
grpcio==1.69.0
grpcio-tools==1.69.0
fastapi==0.115.6pip
uvicorn==0.34.0
kafka-python==2.0.2
pymongo==4.10.1
